requirements.txt
matplotlib==3.5.1
numpy==1.22.0
paho_mqtt==1.6.1
pandas==1.3.5
pytest==7.1.1
tqdm==4.63.0
